+++ src/app/layout/page/overlay/overlay.component.ts
@@ -7,8 +7,8 @@ import { AppUpdateState, FeatureModule, VisibleFlag } from '@app/type';
 import { SnackBarService } from '@shared/module/material/service';
 import { ContextService } from '@shared/module/poe/service';
 import { Context } from '@shared/module/poe/type';
-import { BehaviorSubject, Observable } from 'rxjs';
-import { distinctUntilChanged, flatMap, map, tap } from 'rxjs/operators';
+import { BehaviorSubject, Observable, of, interval, timer, EMPTY } from 'rxjs';
+import { delayWhen, distinctUntilChanged, flatMap, map, tap, debounce } from 'rxjs/operators';
 import { UserSettingsService } from '../../service/user-settings.service';
 import { UserSettings } from '../../type';
 
@@ -101,13 +101,13 @@ export class OverlayComponent implements OnInit, OnDestroy {
     this.app.visibleChange().pipe(
       tap(flag => this.shortcut.check(flag)),
       map(flag => flag !== VisibleFlag.None),
+      debounce(show => show ? EMPTY : timer(1500)),
       distinctUntilChanged()
     ).subscribe(show => {
-      if (!show) {
-        this.window.hide();
-      } else {
+      if (show) {
         this.window.show();
-        this.game.focus();
+      } else {
+        this.window.hide();
       }
     });
     this.app.triggerVisibleChange();